+++ app/src/main/java/woowacourse/shopping/presentation/ui/cart/CartActivity.kt
@@ -9,8 +9,8 @@ import androidx.activity.viewModels
 import woowacourse.shopping.R
 import woowacourse.shopping.databinding.ActivityCartBinding
 import woowacourse.shopping.presentation.base.BindingActivity
-import woowacourse.shopping.presentation.ui.cart.fragment.CartListFragment
-import woowacourse.shopping.presentation.ui.cart.fragment.RecommendFragment
+import woowacourse.shopping.presentation.ui.cart.cartList.CartListFragment
+import woowacourse.shopping.presentation.ui.cart.recommend.RecommendFragment
 import woowacourse.shopping.presentation.ui.model.UpdatedProductData
 import woowacourse.shopping.presentation.ui.shopping.ShoppingActivity
 import woowacourse.shopping.presentation.util.EventObserver
