+++ app/src/main/java/woowacourse/shopping/presentation/ui/cart/cartList/CartViewHolder.kt
@@ -1,4 +1,4 @@
-package woowacourse.shopping.presentation.ui.cart.adapter
+package woowacourse.shopping.presentation.ui.cart.cartList
 
 import androidx.recyclerview.widget.RecyclerView
 import woowacourse.shopping.databinding.ItemCartBinding
